[refactor]ILクラスファイル分割
